requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.11
django-bootstrap4==24.1
django-ckeditor==6.7.1
django-jazzmin==2.6.1
django-js-asset==2.2.0
django-multifilefield==1.0
django-multiupload==0.6.1
django-richtextfield==1.6.2
django-shortcuts==1.6
et-xmlfile==1.1.0
gunicorn==22.0.0
idna==3.7
openpyxl==3.1.2
packaging==24.0
paystack==1.5.0
pillow==10.3.0
pypaystack==1.24
python-dotenv==1.0.1
requests==2.32.2
simplejson==3.19.2
soupsieve==2.5
sqlparse==0.4.4
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
